Move ClassLoader utils to own package
Just some simple and largely stylistic refactoring that was pat of an
abandoned effort to make classloader utils available everywhere.
