+++ src/main/scala/higherkindness/rules_scala/common/classloaders/BUILD
@@ -1,0 +1,14 @@
+load("//rules:scala.bzl", "scala_library")
+load("//rules:scalafmt.bzl", "scala_format_test")
+
+scala_library(
+    name = "classloaders",
+    srcs = glob(["**/*.scala"]),
+    scala = "//src/main/scala:bootstrap",
+    visibility = ["//visibility:public"],
+)
+
+scala_format_test(
+    name = "format",
+    srcs = glob(["**/*.scala"]),
+)
