+++ src/main/scala/higherkindness/rules_scala/common/sbt-testing/SubprocessRunner.scala
@@ -1,6 +1,7 @@
 package higherkindness.rules_scala
 package common.sbt_testing
 
+import common.classloaders.ClassLoaders
 import java.io.ObjectInputStream
 import java.nio.file.Paths
 import scala.collection.mutable
