+++ src/main/scala/higherkindness/rules_scala/workers/zinc/test/BUILD
@@ -8,6 +8,7 @@ scala_library(
     visibility = ["//visibility:public"],
     deps = [
         "//src/main/scala/higherkindness/rules_scala/common/args",
+        "//src/main/scala/higherkindness/rules_scala/common/classloaders",
         "//src/main/scala/higherkindness/rules_scala/common/sbt-testing:common",
         "//src/main/scala/higherkindness/rules_scala/workers/common",
         "@annex//:net_sourceforge_argparse4j_argparse4j",
